Split optional modules
the aexpect package is becoming bigger, it might be useful to split it
across multiple optional modules. Let's start with "remote" and "ops"
modules. The difference is only in way we import those modules, instead
of:

    from aexpect import remote_door

we should do:

    from aexpect_remote import remote_door

although, for compatibility reasons, some of the modules are still
available directly in aexpect, when installed on the system.
