+++ tests/test_remote_door.py
@@ -230,12 +230,12 @@ class RemoteDoorTest(unittest.TestCase):
             self.assertEqual(exception, get_first_arg(register_method.mock_calls[i]))
 
         register_method.reset_mock()
-        preselected_modules = ["aexpect.exceptions", "aexpect.remote"]
+        preselected_modules = ["aexpect.exceptions", "aexpect_remote.remote"]
         remote_door.import_remote_exceptions([], modules=preselected_modules)
         imported_classes = [get_first_arg(c) for c in register_method.mock_calls]
         # assert some detected exceptions from the exceptions module
         self.assertIn("aexpect.exceptions.ExpectError", imported_classes)
         self.assertIn("aexpect.exceptions.ShellError", imported_classes)
         # assert some detected exceptions from the remote module
-        self.assertIn("aexpect.remote.RemoteError", imported_classes)
-        self.assertIn("aexpect.remote.UDPError", imported_classes)
+        self.assertIn("aexpect_remote.remote.RemoteError", imported_classes)
+        self.assertIn("aexpect_remote.remote.UDPError", imported_classes)
